+++ printing/lib/src/pdf_preview.dart
@@ -29,6 +29,7 @@ class PdfPreview extends StatefulWidget {
     this.scrollViewDecoration,
     this.pdfPreviewPageDecoration,
     this.pdfFileName,
+    this.useActions = true,
   }) : super(key: key);
 
   /// Called when a pdf document is needed
@@ -42,6 +43,9 @@ class PdfPreview extends StatefulWidget {
 
   /// Add a button to share the pdf document
   final bool allowSharing;
+
+  /// Allow disable actions
+  final bool useActions;
 
   /// Maximum width of the pdf document on screen
   final double maxPageWidth;
@@ -356,7 +360,7 @@ class _PdfPreviewState extends State<PdfPreview> {
       mainAxisAlignment: MainAxisAlignment.center,
       children: <Widget>[
         Expanded(child: scrollView),
-        if (actions.isNotEmpty)
+        if (actions.isNotEmpty && widget.useActions)
           Material(
             elevation: 4,
             color: theme.primaryColor,